backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.1.1
pydantic==2.11.7

# LiveKit dependencies
livekit==1.0.13
livekit-agents==1.2.9
livekit-api==1.0.5
livekit-protocol==1.0.6

# AI and ML dependencies
google-generativeai==0.8.5
mem0ai==0.1.117
transformers==4.56.1
sentence-transformers==5.1.0
torch==2.8.0
huggingface-hub==0.35.0

# Vector database
qdrant-client==1.15.1

# Core Python utilities
aiohttp==3.12.15
asyncio
requests==2.32.5
numpy==2.3.3
pandas==2.3.2

# Authentication and security
PyJWT==2.10.1
cryptography==45.0.7

# Development tools (optional - remove if not needed)
# black==25.1.0
# pytest==8.4.2